+++ src/main/java/io/dataspaceconnector/services/ids/ConnectorService.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2020 Fraunhofer Institute for Software and Systems Engineering
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.dataspaceconnector.services.ids;
 
 import de.fraunhofer.iais.eis.*;
@@ -168,10 +183,9 @@ public class ConnectorService {
      * @return The ids resource.
      */
     public Resource getOfferedResourceById(final URI resourceId) {
-        // TODO The filter not working
         final var resource = offeredResourceService.getAll(Pageable.unpaged())
                 .stream()
-                .filter(x -> x.getId().toString().contains(resourceId.toString()))
+                .filter(x -> resourceId.toString().contains(x.getId().toString()))
                 .findAny();
 
         return resource.map(resourceBuilder::create).orElse(null);