+++ src/main/java/io/dataspaceconnector/services/ids/builder/IdsResourceBuilder.java
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2020 Fraunhofer Institute for Software and Systems Engineering
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.dataspaceconnector.services.ids.builder;
 
 import de.fraunhofer.iais.eis.ConnectorEndpointBuilder;
@@ -67,6 +82,12 @@ public final class IdsResourceBuilder<T extends Resource>
                 ._accessURL_(selfLink)
                 ._endpointDocumentation_(Util.asList(endpointDocs))
                 .build();
+
+        // Build resource only if at least one representation and one contract is present.
+        if (representations.isEmpty() || contracts.isEmpty() || representations.get().isEmpty()
+                || contracts.get().isEmpty()) {
+            return null;
+        }
 
         final var builder = new ResourceBuilder(selfLink)
 //                ._accrualPeriodicity_()