+++ src/test/java/io/dataspaceconnector/services/ids/IdsContractBuilderTest.java
@@ -1,8 +1,29 @@
+/*
+ * Copyright 2020 Fraunhofer Institute for Software and Systems Engineering
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.dataspaceconnector.services.ids;
 
 import de.fraunhofer.iais.eis.Action;
 import de.fraunhofer.isst.ids.framework.configuration.SerializerProvider;
-import io.dataspaceconnector.model.*;
+import io.dataspaceconnector.model.AbstractEntity;
+import io.dataspaceconnector.model.Contract;
+import io.dataspaceconnector.model.ContractDesc;
+import io.dataspaceconnector.model.ContractFactory;
+import io.dataspaceconnector.model.ContractRule;
+import io.dataspaceconnector.model.ContractRuleDesc;
+import io.dataspaceconnector.model.ContractRuleFactory;
 import io.dataspaceconnector.services.ids.builder.IdsContractBuilder;
 import io.dataspaceconnector.services.ids.builder.IdsDutyBuilder;
 import io.dataspaceconnector.services.ids.builder.IdsPermissionBuilder;
@@ -20,7 +41,11 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.UUID;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest(classes = {ContractFactory.class, ContractRuleFactory.class,
         IdsContractBuilder.class, IdsPermissionBuilder.class, IdsProhibitionBuilder.class,
@@ -102,7 +127,7 @@ public class IdsContractBuilderTest {
     }
 
     @Test
-    public void create_maxDepth0_returnContractWithoutRules() {
+    public void create_maxDepth0_returnNull() {
         /* ARRANGE */
         final var contract = getContract();
 
@@ -110,18 +135,7 @@ public class IdsContractBuilderTest {
         final var idsContract = idsContractBuilder.create(contract, 0);
 
         /* ASSERT */
-        assertTrue(idsContract.getId().isAbsolute());
-        assertTrue(idsContract.getId().toString().contains(contract.getId().toString()));
-
-        assertEquals(provider, idsContract.getProvider());
-        assertEquals(consumer, idsContract.getConsumer());
-        assertEquals(IdsUtils.getGregorianOf(date), idsContract.getContractStart());
-        assertEquals(IdsUtils.getGregorianOf(date), idsContract.getContractEnd());
-        assertNull(idsContract.getProperties());
-
-        assertNull(idsContract.getPermission());
-        assertNull(idsContract.getProhibition());
-        assertNull(idsContract.getObligation());
+        assertNull(idsContract);
     }
 
     @Test