+++ src/test/java/io/dataspaceconnector/services/ids/IdsResourceBuilderTest.java
@@ -1,9 +1,46 @@
+/*
+ * Copyright 2020 Fraunhofer Institute for Software and Systems Engineering
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package io.dataspaceconnector.services.ids;
 
 import de.fraunhofer.iais.eis.Language;
 import de.fraunhofer.isst.ids.framework.configuration.SerializerProvider;
-import io.dataspaceconnector.model.*;
-import io.dataspaceconnector.services.ids.builder.*;
+import io.dataspaceconnector.model.AbstractEntity;
+import io.dataspaceconnector.model.Artifact;
+import io.dataspaceconnector.model.ArtifactDesc;
+import io.dataspaceconnector.model.ArtifactFactory;
+import io.dataspaceconnector.model.Contract;
+import io.dataspaceconnector.model.ContractDesc;
+import io.dataspaceconnector.model.ContractFactory;
+import io.dataspaceconnector.model.ContractRule;
+import io.dataspaceconnector.model.ContractRuleDesc;
+import io.dataspaceconnector.model.ContractRuleFactory;
+import io.dataspaceconnector.model.OfferedResource;
+import io.dataspaceconnector.model.OfferedResourceDesc;
+import io.dataspaceconnector.model.OfferedResourceFactory;
+import io.dataspaceconnector.model.Representation;
+import io.dataspaceconnector.model.RepresentationDesc;
+import io.dataspaceconnector.model.RepresentationFactory;
+import io.dataspaceconnector.model.Resource;
+import io.dataspaceconnector.services.ids.builder.IdsArtifactBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsContractBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsDutyBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsPermissionBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsProhibitionBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsRepresentationBuilder;
+import io.dataspaceconnector.services.ids.builder.IdsResourceBuilder;
 import lombok.SneakyThrows;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,7 +53,11 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.UUID;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 @SpringBootTest(classes = {OfferedResourceFactory.class, RepresentationFactory.class,
         ArtifactFactory.class, ContractFactory.class, ContractRuleFactory.class,
@@ -139,7 +180,7 @@ public class IdsResourceBuilderTest {
     }
 
     @Test
-    public void create_maxDepth0_returnResourceWithoutRepresentationsAndContracts() {
+    public void create_maxDepth0_returnNull() {
         /* ARRANGE */
         final var resource = getOfferedResource();
 
@@ -147,26 +188,11 @@ public class IdsResourceBuilderTest {
         final var idsResource = idsResourceBuilder.create(resource, 0);
 
         /* ASSERT */
-        assertTrue(idsResource.getId().isAbsolute());
-        assertTrue(idsResource.getId().toString().contains(idsResource.getId().toString()));
-
-        assertEquals(publisher, idsResource.getPublisher());
-        assertEquals(sovereign, idsResource.getSovereign());
-        assertEquals(license, idsResource.getStandardLicense());
-        assertEquals(1, idsResource.getDescription().size());
-        assertEquals(description, idsResource.getDescription().get(0).getValue());
-        assertEquals(1, idsResource.getKeyword().size());
-        assertEquals(keyword, idsResource.getKeyword().get(0).getValue());
-        assertEquals(1, idsResource.getTitle().size());
-        assertEquals(title, idsResource.getTitle().get(0).getValue());
-        assertNull(idsResource.getProperties());
-
-        assertNull(idsResource.getRepresentation());
-        assertNull(idsResource.getContractOffer());
+        assertNull(idsResource);
     }
 
     @Test
-    public void create_maxDepth1_returnResourceWithoutArtifactsAndRules() {
+    public void create_maxDepth1_returnNull() {
         /* ARRANGE */
         final var resource = getOfferedResource();
 
@@ -174,29 +200,7 @@ public class IdsResourceBuilderTest {
         final var idsResource = idsResourceBuilder.create(resource, 1);
 
         /* ASSERT */
-        assertTrue(idsResource.getId().isAbsolute());
-        assertTrue(idsResource.getId().toString().contains(idsResource.getId().toString()));
-
-        assertEquals(publisher, idsResource.getPublisher());
-        assertEquals(sovereign, idsResource.getSovereign());
-        assertEquals(license, idsResource.getStandardLicense());
-        assertEquals(1, idsResource.getDescription().size());
-        assertEquals(description, idsResource.getDescription().get(0).getValue());
-        assertEquals(1, idsResource.getKeyword().size());
-        assertEquals(keyword, idsResource.getKeyword().get(0).getValue());
-        assertEquals(1, idsResource.getTitle().size());
-        assertEquals(title, idsResource.getTitle().get(0).getValue());
-        assertNull(idsResource.getProperties());
-
-        final var representations = idsResource.getRepresentation();
-        assertEquals(1, representations.size());
-        assertNull(representations.get(0).getInstance());
-
-        final var contracts = idsResource.getContractOffer();
-        assertEquals(1, contracts.size());
-        assertNull(contracts.get(0).getPermission());
-        assertNull(contracts.get(0).getProhibition());
-        assertNull(contracts.get(0).getProhibition());
+        assertNull(idsResource);
     }
 
     @Test